Added warning message for server scripts that might be causing a long freeze. Applied a couple of minor source tweaks.

+++ MTA10/mods/deathmatch/logic/CPacketHandler.cpp
@@ -4796,7 +4796,7 @@ void CPacketHandler::Packet_ResourceStart ( NetBitStreamInterface& bitStream )
                     bitStream.Read ( (char*)chunkChecksum.md5.data, sizeof ( chunkChecksum.md5.data ) );
                     bitStream.Read ( dChunkDataSize );
 
-                    uiTotalSizeProcessed += dChunkDataSize;
+                    uiTotalSizeProcessed += (uint)dChunkDataSize;
                     if ( uiTotalSizeProcessed / 1024 / 1024 > 50 )
                         g_pCore->UpdateDummyProgress( uiTotalSizeProcessed / 1024 / 1024, " MB" );
